agents/requirements.txt
google-adk>=1.26.0
google-genai>=1.65.0
httpx>=0.28.0
